pyproject.toml
[tool.poetry]
name = "whatsupyoutube"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
ipykernel = "^6.24.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.93.0"

[tool.poetry.group.dev.dependencies]
pytest = "^6.0.0"
pytest-mock = "*"
black = "^23.7.0"
mypy = "^1.4.1"
isort = "^5.12.0"
pydocstyle = "^6.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
